refactor(navigation): reorganize protected routes and remove unused screens
- Add animalsList and registerAnimal screens to protected layout
- Remove standalone animalsList, registerAnimal, and animal/[animalId] screens
